Initial commit.
The MICROS buildassistant is an all-in-one package for playing around
MICROS. This docker image can be run easily and will start a run of
the OS all without installing anyting (except docker).

This first commit is the initial scenario of just launching the
image and looking at the OS booting up. It does not provide
rebuilding mechanisms or any persistance. I might extent this image
or make new docker images to allow more options.